+++ canvas-design-tools/canvas-design-tools/helpers/paths.js
@@ -18,8 +18,13 @@ const packageFolder = path.resolve(rootFolder, 'node_modules', '@we-make-website
  */
 module.exports = {
   config: {
-    internal: path.resolve(packageFolder, 'helpers', 'design-config.js'),
-    project: path.resolve(rootFolder, 'canvas', 'design-config.js'),
+    internal: path.resolve(packageFolder, 'design.config.js'),
+    project: {
+      // Order is important
+      root: path.resolve(rootFolder, 'design.config.js'),
+      canvas: path.resolve(rootFolder, 'canvas', 'design-config.js'),
+      helpers: path.resolve(rootFolder, 'canvas', 'helpers', 'design-config.js'),
+    },
   },
   icons: path.resolve(rootFolder, 'src', 'icons'),
   scripts: {
@@ -28,6 +33,7 @@ module.exports = {
   styles: {
     base: path.resolve(rootFolder, 'src', 'styles', 'base'),
     config: path.resolve(rootFolder, 'src', 'styles', 'config'),
+    helpers: path.resolve(rootFolder, 'src', 'styles', 'helpers'),
   },
   storybook: {
     assets: path.resolve(rootFolder, '.storybook', 'assets'),