+++ canvas-component-tools/CHANGELOG.md
@@ -1,6 +1,28 @@
-# 📅 Canvas Component Tools Changelog
+# 📚 Canvas Component Tools Changelog
 
 Canvas Component Tools uses [semantic versioning](https://semver.org/).
+
+## 1.3.0 - 2023-04-11
+
+* Added `disabled_on` property to async schema files to disable in header and footer section groups by default
+
+## 1.2.1 - 2023-03-01
+
+* Fixed limited interactivity Vue template failing eslint
+
+## 1.2.0 - 2023-03-01
+
+* Updated limited interactivity template to include class in template
+
+## 1.1.0 - 2023-01-31
+
+* Added limited interactivity to template options
+* Added documentation links to footer as choices are made
+* Updated template titles to not use 404 title to avoid confusion
+* Updated command to output version of the tool instead of Canvas
+* Updated dependencies
+* Removed `presets` array from global schema template
+* Fixed path to `getCanvasConfig()` helper function
 
 ## 1.0.0 - 2022-11-03