+++ canvas-design-tools/CHANGELOG.md
@@ -2,6 +2,41 @@
 
 Canvas Design Tools uses [semantic versioning](https://semver.org/).
 
+## 2.0.0 - 2023-04-11
+
+* Added `-hover` text classes
+* Updated to support new design tokens
+* Updated forms styleguide template
+
+## 1.4.0 - 2023-02-28
+
+* Added support for utility stylesheets
+* Added `--no-js` and `--no-storybook` flags
+
+## 1.3.1 - 2023-02-01
+
+* Fixed Storybook text styles being affected by tokens
+* Fixed `include` array setting in critical stylesheet overriding default text classes
+
+## 1.3.0 - 2023-01-31
+
+* Updated command to output version of the tool instead of Canvas
+* Updated dependencies
+* Fixed token names starting with numbers causing issues in breakpoints and mq-breakpoints templates
+
+## 1.2.0 - 2022-11-17
+
+* Added support for all previous project locations of _design.config.js_
+* Updated so that typography classes are sorted based on `sorting` config
+* Fixed default config not supporting paragraph indent tokens
+* Fixed design config trying to use Basis paths helper
+
+## 1.1.0 - 2022-11-16
+
+* Added support for default styles using the same class
+* Updated defaults config to use class names to match instead of descriptions
+* Updated to automatically add `line-height.baseline` token if missing
+
 ## 1.0.0 - 2022-11-03
 
 * Initial release